requirements.txt
# Core dependencies for Gait Biometric Identification System
# Python >= 3.8 required

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.62.0

# Utilities
h5py>=3.6.0
joblib>=1.2.0
tqdm>=4.62.0

